# openapi/swagger.json
{
  "x-generator": "NSwag v13.5.0.0 (NJsonSchema v10.1.15.0 (Newtonsoft.Json v11.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:58765",
  "schemes": [
    "http"
  ],
  "paths": {
    "/Label/delete/{identcode}": {
      "delete": {
        "tags": [
          "Label"
        ],
        "operationId": "Label_DeleteLabel",
        "parameters": [
          {
            "type": "string",
            "name": "identcode",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/Label/label/{identcode}": {
      "get": {
        "tags": [
          "Label"
        ],
        "operationId": "Label_GetLabel",
        "parameters": [
          {
            "type": "string",
            "name": "identcode",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/Label/label": {
      "post": {
        "tags": [
          "Label"
        ],
        "operationId": "Label_CreateLabel",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "label",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LabelDTO"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "LabelDTO": {
      "allOf": [
        {
          "$ref": "#/definitions/BaseLabelOfAddressDTOAndFeaturesDTO"
        },
        {
          "type": "object",
          "required": [
            "sender",
            "receiver",
            "features"
          ],
          "properties": {
            "sender": {
              "$ref": "#/definitions/AddressDTO"
            },
            "receiver": {
              "$ref": "#/definitions/AddressDTO"
            },
            "features": {
              "$ref": "#/definitions/FeaturesDTO"
            }
          }
        }
      ]
    },
    "AddressDTO": {
      "type": "object",
      "required": [
        "name",
        "street",
        "homeNo",
        "zip",
        "country",
        "city"
      ],
      "properties": {
        "company": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "minLength": 1
        },
        "surname": {
          "type": "string"
        },
        "street": {
          "type": "string",
          "minLength": 1
        },
        "homeNo": {
          "type": "string",
          "minLength": 1
        },
        "zip": {
          "type": "string",
          "minLength": 1
        },
        "country": {
          "type": "string",
          "minLength": 1
        },
        "city": {
          "type": "string",
          "minLength": 1
        },
        "email": {
          "type": "string"
        },
        "mobile": {
          "type": "string"
        }
      }
    },
    "FeaturesDTO": {
      "type": "object",
      "required": [
        "price",
        "currency",
        "weight"
      ],
      "properties": {
        "deliveryIns": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "decimal"
        },
        "currency": {
          "type": "string",
          "minLength": 1
        },
        "weight": {
          "type": "number",
          "format": "decimal"
        }
      }
    },
    "BaseLabelOfAddressDTOAndFeaturesDTO": {
      "type": "object",
      "x-abstract": true,
      "properties": {
        "sender": {
          "$ref": "#/definitions/AddressDTO"
        },
        "receiver": {
          "$ref": "#/definitions/AddressDTO"
        },
        "features": {
          "$ref": "#/definitions/FeaturesDTO"
        }
      }
    }
  }
}